Refactor: Function playCpuSequence refactored, playSound function added.
playCpuSequence refactored by removing switch statement to change colors.
Decision of whose button to change color is made based on the id of the clicked button.
This is done with the help of the added object colorCodes which stores the color codes. id of the clicked button is used as the key.
playSound function still uses switch statements and plays the corresponding sound based on the provided id.
playSound function will be used for both cpu and player.