Feat: Basic game functionality implemented.
Now CPU plays its sequence and waits for the player sequence.
The player sequence is compared to CPU's sequence and score is incremented if the sequence is right.
Added a condition at the beginning of mouse event handlers to prevent user from giving any input if the CPU sequence is playing.
At this point game doesn't stops if the user's input is wrong but doesn't restart the game.
Game goes on indefinitely.