Feat: No strict mode implemented.
When the player play the sequence incorrectly, the new game plays the same old sequence if strict mode is off.
The change was made in the generateSequence function by adding if statement that checks for strict mode.
If strict mode is off, sequence is copied from noStrictSequence which is permanently saved.
If strict mode is on, sequence is randomly generated.
Two variables added, strictStatus and noStrictSequence.