Bug: Fixed a bug with returning to no strict mode.
The default of the game is no strict mode.
The bug was that the player wouldn't be able to go back to normal mode from strict mode.
The buf is fixed by inverting the strictClicked boolean in the strict-button handler and an additional if statment in the start button handler.