Refactor: The functionality of strict and normal modes modified.
I got the rules of the game incorrectly before.
Now, with strict mode off, the game doesn't restart when the user makes a mistake.
A message warning appears telling the player that the sequence was wrong and the player can try again.
With strict mode on, if the player makes a mistake the game ends.
When a new game is started, a new random sequence is played.
Removed variables are: round, noStrictSequence, strictStatus.
Added functions: continue button handler.
Modified functions: color-button mouseup handler, generateSequence.